requirements.txt
# AG Futures Production - Python Dependencies
# Core data manipulation
pandas>=2.2.0
numpy>=1.24.0
python-dateutil>=2.9.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Data collection
yfinance>=0.2.40
requests>=2.31.0
beautifulsoup4>=4.12.2

# Optional: API framework (if deploying REST API)
# fastapi>=0.104.0
# uvicorn>=0.24.0
# pydantic>=2.4.0
